backend/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.1
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
drf-spectacular==0.27.2
